// apps/web/src/lib/api/generated/types/getCandidatesPlacementStatusesItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type GetCandidatesPlacementStatusesItem = typeof GetCandidatesPlacementStatusesItem[keyof typeof GetCandidatesPlacementStatusesItem];


export const GetCandidatesPlacementStatusesItem = {
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
} as const;