// apps/web/src/lib/api/generated/types/getCandidatesSubmissionStatusesItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type GetCandidatesSubmissionStatusesItem = typeof GetCandidatesSubmissionStatusesItem[keyof typeof GetCandidatesSubmissionStatusesItem];


export const GetCandidatesSubmissionStatusesItem = {
  SUBMITTED: 'SUBMITTED',
  INTERVIEWING: 'INTERVIEWING',
  REJECTED: 'REJECTED',
  PLACED: 'PLACED',
} as const;